use chakra Card component
